+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## To Be Released
 
+* feat(private-networks): add command to list private network domain names
+
 ## 1.39.0
 
 * task(backups-config): update error message when unscheduling periodic backup